requirements.txt
twitchio==2.10.0
aiogram==3.4.1
PyYAML==6.0.1
aiosqlite==0.20.0
